pyproject.toml
[project]
name = "ttring"
version = "0.1.0"
description = "Rings the bell"
requires-python = ">=3.9"
dependencies = [
    "mysql-connector-python>=9.1.0",
    "rich>=13.9.4",
    "ruff>=0.8.4",
    "survey>=5.4.2",
    "tabulate>=0.9.0",
]
